requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.6.3
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
Django==2.1.1
djangorestframework==3.8.2
idna==2.7
lxml==4.2.5
mysqlclient==1.3.13
numpy==1.15.0
Pillow==5.2.0
pycparser==2.18
pymongo==3.7.1
PyMySQL==0.9.2
pytz==2018.5
requests==2.19.1
selenium==3.14.1
six==1.11.0
urllib3==1.23
virtualenv==16.0.0
